requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.1
Django==3.2.5
django-bootstrap4==2.0.1
django-filter==2.3.0
djangorestframework
Markdown
Pillow
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
setuptools
